Move unintuitive test/test.js helper and document it
