+++ lib/build.test.js
@@ -1,4 +1,4 @@
-const { expect, log, fixturePath } = require('../test/test');
+const { expect, log, fixturePath } = require('../test');
 const { Builder } = require('./build');
 
 const sinon = require('sinon');
