+++ lib/logger.test.js
@@ -1,7 +1,7 @@
-// Warning: This test doesn't work because Logger is completely mocked out in ./test/test.js
+// Warning: This test doesn't work because Logger is completely mocked out in ../test/index.js
 // Ideally we'd change this so we can make assertions against the class itself
 
-// const { expect, sinon } = require('../test/test');
+// const { expect, sinon } = require('../test');
 // const { Logger, LogLevel } = require('./logger');
 // const chalk = require('chalk');
