+++ lib/platform.test.js
@@ -1,5 +1,5 @@
 const sinon = require('sinon');
-const { expect } = require('../test/test');
+const { expect } = require('../test');
 const { PLATFORMS, Platform, platformForName } = require('./platform');
 
 describe('platform.js', () => {
