+++ lib/runner.test.js
@@ -1,4 +1,4 @@
-const { expect, config, log, fixturePath } = require('../test/test');
+const { expect, config, log, fixturePath } = require('../test');
 const proxyquire = require('proxyquire');
 const { Runner } = proxyquire('../lib/runner', { './config': { config } });
 const { RunMode } = require('./config');
